+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 <!--
-  Copyright (c) 2023 Michael Federczuk
+  Copyright (c) 2025 Michael Federczuk
   SPDX-License-Identifier: CC-BY-SA-4.0
 -->
 
@@ -10,6 +10,14 @@
 All notable changes to this project will be documented in this file.
 The format is based on [**Keep a Changelog v1.0.0**](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [**Semantic Versioning v2.0.0**](https://semver.org/spec/v2.0.0.html).
+
+## Unreleased ##
+
+### Deprecated ###
+
+* The type [`GenericKey`](src/types.ts) is deprecated in favor of the built-in type `PropertyKey`.  
+  `PropertyKey` is identical to the previous definition of `GenericKey`. `GenericKey` now is simply an alias for
+  `PropertyKey`
 
 ## [v2.0.0-rc01] - 2023-01-29 ##